Added Window management for graphs with Graph image management
If the user clicks statistics, then he can now view a new window, this window has separate tabs, which relate to the topic of the grahps. There are scrollbars on the tabs so you can basically end an endless amount of graphs in the 'views' tab and so on. The script Graph.py runs and creates the graph, and exports it to Assets/Graphs/. From here, main.py takes it and displays it in the statistics window.